Disable copy a bookmark if the edited bookmark is not the current.
This fixes case when the edit form is invoked for the not current bookmark
and checking copy will copy the CURRENT bookmark but not edited one.
To copy not current bookmark it will be needed to clone the request
object but it's too complicated to implement. So copy is actual only
for the current bookmark.